+++ java-service-processor/src/test/resources/definition/TestFilterCheckedException.java
@@ -1,0 +1,13 @@
+package definition;
+
+import nbbrd.service.ServiceDefinition;
+import nbbrd.service.ServiceFilter;
+
+import java.io.IOException;
+
+@ServiceDefinition
+interface TestFilterCheckedException {
+
+    @ServiceFilter
+    boolean isAvailable() throws IOException;
+}