+++ java-service-processor/src/test/resources/definition/TestIdCheckedException.java
@@ -1,0 +1,13 @@
+package definition;
+
+import nbbrd.service.ServiceDefinition;
+import nbbrd.service.ServiceId;
+
+import java.io.IOException;
+
+@ServiceDefinition
+interface TestIdCheckedException {
+
+    @ServiceId
+    String getName() throws IOException;
+}