+++ java-service-processor/src/test/resources/definition/TestSorterCheckedException.java
@@ -1,0 +1,13 @@
+package definition;
+
+import nbbrd.service.ServiceDefinition;
+import nbbrd.service.ServiceSorter;
+
+import java.io.IOException;
+
+@ServiceDefinition
+interface TestSorterCheckedException {
+
+    @ServiceSorter
+    int getCost() throws IOException;
+}